+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # clever-tools changelog
 
-## Unrelease (????-??-??)
+## Unreleased (????-??-??)
+
+* add config file and auth source to `clever diag`
+* add shell to `clever diag`
+* improve `clever diag`, display (color and details) for oAuth token and user ID
 
 ## 2.11.0 (2023-07-25)